Optimize APIGroup E2E test case
